+++ src/EliteChroma.Core/Layers/FssLayer.cs
@@ -1,6 +1,5 @@
 ﻿using System.Collections.Generic;
 using System.Diagnostics.CodeAnalysis;
-using Colore.Data;
 using Colore.Effects.Keyboard;
 using EliteChroma.Chroma;
 using EliteFiles.Bindings.Binds;
@@ -11,9 +10,6 @@ namespace EliteChroma.Core.Layers
     [SuppressMessage("Performance", "CA1812:Avoid uninstantiated internal classes", Justification = "Instantiated by ChromaController.InitChromaEffect().")]
     internal sealed class FssLayer : LayerBase
     {
-        private static readonly Color EliteOrange = new Color(1.0, 0.2, 0);
-        private static readonly Color AnalysisModeColor = new Color(0.14, 0.62, 0.81);
-
         private static readonly IReadOnlyCollection<string> FssCamera = new[]
         {
             FullSpectrumSystemScanner.CameraPitch,
@@ -46,15 +42,15 @@ namespace EliteChroma.Core.Layers
             }
 
             var k = canvas.Keyboard;
-            k[Key.Escape] = Color.White;
+            k[Key.Escape] = Colors.InterfaceMode;
 
-            ApplyColorToBinding(canvas.Keyboard, FssCamera, Color.Green);
-            ApplyColorToBinding(canvas.Keyboard, FssZoom, Color.Blue);
-            ApplyColorToBinding(canvas.Keyboard, FssTuning, Color.Purple);
-            ApplyColorToBinding(canvas.Keyboard, FullSpectrumSystemScanner.DiscoveryScan, AnalysisModeColor);
-            ApplyColorToBinding(canvas.Keyboard, FullSpectrumSystemScanner.Quit, Color.White);
-            ApplyColorToBinding(canvas.Keyboard, FullSpectrumSystemScanner.Target, Color.Yellow);
-            ApplyColorToBinding(canvas.Keyboard, FullSpectrumSystemScanner.ShowHelp, EliteOrange);
+            ApplyColorToBinding(canvas.Keyboard, FssCamera, Colors.FssCamera);
+            ApplyColorToBinding(canvas.Keyboard, FssZoom, Colors.FssZoom);
+            ApplyColorToBinding(canvas.Keyboard, FssTuning, Colors.FssTuning);
+            ApplyColorToBinding(canvas.Keyboard, FullSpectrumSystemScanner.DiscoveryScan, Game.Colors.AnalysisMode);
+            ApplyColorToBinding(canvas.Keyboard, FullSpectrumSystemScanner.Quit, Colors.InterfaceMode);
+            ApplyColorToBinding(canvas.Keyboard, FullSpectrumSystemScanner.Target, Colors.FssTarget);
+            ApplyColorToBinding(canvas.Keyboard, FullSpectrumSystemScanner.ShowHelp, Game.Colors.Hud);
         }
     }
 }