+++ src/EliteChroma.Core/Layers/HyperspaceLayer.ParticlePopulation.cs
@@ -1,5 +1,6 @@
 ﻿using System.Collections.Generic;
 using Colore.Data;
+using EliteChroma.Elite;
 
 namespace EliteChroma.Core.Layers
 {
@@ -9,17 +10,17 @@ namespace EliteChroma.Core.Layers
         {
             public static readonly IReadOnlyCollection<ParticlePopulation> Supercruise = new[]
             {
-                new ParticlePopulation(JumpPhase.PreJump, 20, -20, Color.Blue),
-                new ParticlePopulation(JumpPhase.Tunnel, 20, -20, Color.Blue),
-                new ParticlePopulation(JumpPhase.Jump, 20, -20, Color.Blue),
+                new ParticlePopulation(JumpPhase.PreJump, 20, -20, GameColors.SupercruiseTunnelParticle),
+                new ParticlePopulation(JumpPhase.Tunnel, 20, -20, GameColors.SupercruiseTunnelParticle),
+                new ParticlePopulation(JumpPhase.Jump, 20, -20, GameColors.SupercruiseTunnelParticle),
             };
 
             public static readonly IReadOnlyCollection<ParticlePopulation> Hyperspace = new[]
             {
-                new ParticlePopulation(JumpPhase.PreJump, 20, -20, Color.Blue),
-                new ParticlePopulation(JumpPhase.Tunnel, 100, -80, Color.White),
-                new ParticlePopulation(JumpPhase.Jump, 10, -20, Color.Blue),
-                new ParticlePopulation(JumpPhase.Jump, 5, -10, new Color(160, 255, 255)),
+                new ParticlePopulation(JumpPhase.PreJump, 20, -20, GameColors.SupercruiseTunnelParticle),
+                new ParticlePopulation(JumpPhase.Tunnel, 100, -80, GameColors.HyperspaceTunnelParticle),
+                new ParticlePopulation(JumpPhase.Jump, 10, -20, GameColors.HyperspaceDimStar),
+                new ParticlePopulation(JumpPhase.Jump, 5, -10, GameColors.HyperspaceBrightStar),
             };
 
             private ParticlePopulation(JumpPhase jumpPhase, double spawnsPerSecond, double zSpeedPerSecond, Color color)