+++ src/EliteChroma.Core/Layers/SrvAnalysisModeLayer.cs
@@ -10,9 +10,6 @@ namespace EliteChroma.Core.Layers
     [SuppressMessage("Performance", "CA1812:Avoid uninstantiated internal classes", Justification = "Instantiated by ChromaController.InitChromaEffect().")]
     internal sealed class SrvAnalysisModeLayer : LayerBase
     {
-        private static readonly Color AnalysisModeColor = new Color(0.14, 0.62, 0.81);
-        private static readonly Color CombatModeColor = Color.Red;
-
         public override int Order => 600;
 
         protected override void OnRender(ChromaCanvas canvas)
@@ -22,35 +19,41 @@ namespace EliteChroma.Core.Layers
                 return;
             }
 
-            var color = Game.Status.HasFlag(Flags.HudInAnalysisMode) ? AnalysisModeColor.Transform(Game.GuiColour) : CombatModeColor;
+            var colorOn = Game.Status.HasFlag(Flags.HudInAnalysisMode) ? Game.Colors.AnalysisMode : Colors.HardpointsToggle;
 
-            ApplyColorToBinding(canvas.Keyboard, DrivingModeSwitches.PlayerHUDModeToggle, color);
+            ApplyColorToBinding(canvas.Keyboard, DrivingModeSwitches.PlayerHUDModeToggle, colorOn);
 
-            ApplyColorToBinding(canvas.Keyboard, Driving.PrimaryFire, color);
-            ApplyColorToBinding(canvas.Keyboard, Driving.SecondaryFire, color);
-            ApplyColorToBinding(canvas.Keyboard, Driving.CycleFireGroupNext, color);
-            ApplyColorToBinding(canvas.Keyboard, Driving.CycleFireGroupPrevious, color);
+            ApplyColorToBinding(canvas.Keyboard, Driving.PrimaryFire, colorOn);
+            ApplyColorToBinding(canvas.Keyboard, Driving.SecondaryFire, colorOn);
+            ApplyColorToBinding(canvas.Keyboard, Driving.CycleFireGroupNext, colorOn);
+            ApplyColorToBinding(canvas.Keyboard, Driving.CycleFireGroupPrevious, colorOn);
 
-            Color hColor = color;
-            Color mColor = color;
-            Color bColor = color.Transform(0.5);
+            var hardpointsDeployed = !Game.Status.HasFlag(Flags.SrvTurretRetracted);
 
-            if (!Game.Status.HasFlag(Flags.SrvTurretRetracted))
+            Color hColor;
+            if (hardpointsDeployed)
             {
                 StartAnimation();
-                hColor = PulseColor(Color.Black, color, TimeSpan.FromSeconds(1));
+                hColor = PulseColor(Color.Black, colorOn, TimeSpan.FromSeconds(1));
             }
             else
             {
                 StopAnimation();
-                mColor = bColor;
+                hColor = colorOn;
             }
 
             ApplyColorToBinding(canvas.Keyboard, Driving.ToggleTurret, hColor);
-            canvas.Mouse.Set(mColor);
-            canvas.Mousepad.Set(mColor);
-            canvas.Headset.Set(mColor);
-            canvas.ChromaLink.Set(bColor);
+
+            var colorOff = Game.Status.HasFlag(Flags.HudInAnalysisMode)
+                ? Game.Colors.AnalysisMode.Transform(Colors.DeviceDimBrightness)
+                : Game.Colors.Hud.Transform(Colors.DeviceDimBrightness);
+
+            var c = hardpointsDeployed ? colorOn : colorOff;
+
+            canvas.Mouse.Set(c);
+            canvas.Mousepad.Set(c);
+            canvas.Headset.Set(c);
+            canvas.ChromaLink.Set(c);
         }
     }
 }