+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
+## [1.8.0](https://github.com/poveden/EliteChroma/compare/v1.7.0...v1.8.0) — 2020-06-23
+
+### Changed
+
+- Apply different background brightnesses per device
+- Change coloring strategy for analysis mode
+
 ## [1.7.0](https://github.com/poveden/EliteChroma/compare/v1.6.0...v1.7.0) — 2020-06-05
 
 ### Fixed